backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.6.0
pydantic-settings==2.1.0

# Google Cloud Vertex AI
google-cloud-aiplatform==1.40.0
google-auth==2.27.0

# OpenAI
openai>=1.30.0

# Vector Database
chromadb==0.4.22
# Pinning onnxruntime to avoid conflicts on Windows
onnxruntime==1.16.3
overrides==7.4.0

# Text Processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
nltk==3.8.1
spacy==3.7.2
tiktoken==0.5.2
unstructured[all-docs]>=0.10.0  # Text extraction and cleaning
ftfy>=6.1.0  # Unicode fixing

# Encryption
cryptography==42.0.2

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1

# Data Science & Visualization
scikit-learn==1.4.0
numpy==1.26.3

# Embedding Providers
sentence-transformers==2.3.1
boto3==1.34.34
